packages/web/backend/pyproject.toml
[project]
name = "opentools-web"
version = "0.1.0"
description = "Web dashboard for OpenTools security toolkit"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "fastapi-users[sqlalchemy]>=13.0",
    "sse-starlette>=2.0",
    "slowapi>=0.1",
    "python-multipart>=0.0.18",
    "bcrypt>=4.0",
    "pydantic-settings>=2.0",
    "opentools",
]

[project.optional-dependencies]
dev = ["pytest>=9.0", "httpx>=0.28", "pytest-asyncio>=0.25", "aiosqlite>=0.19"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
